requirements.txt
click>=7.1.2
click_repl
verboselogs>=1.7
pexpect>=4.8.0
pyserial>=3.4
ruamel.yaml
setuptools>=46.1.3
psutil>=5.7.0
PyYAML>=5.3.1